requirements.txt
click==8.0.1
dataclasses==0.8
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pkg-resources==0.0.0
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
